refactor(chat): share SSE reader across proxy and direct streams
streamChat had the same drain/split/slice(6)/dispatch loop copy-pasted
twice within one function (main read and final-flush passes), and
stream-direct.ts carried a third near-identical copy. Promote the
readSseStream helper (formerly private in stream-direct.ts) into
stream-client.ts, export it, and rewrite streamChat to use it. Net
removal of ~65 lines; behavior unchanged — existing stream-client and
stream-direct tests continue to pass.
